% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_levels.R
\name{ext_levels}
\alias{ext_levels}
\title{Enter a vector and return a character vector of the elements with duplicates removed.}
\usage{
ext_levels(vector)
}
\arguments{
\item{vector}{target vector}
}
\description{
Enter a vector and return a character vector of the elements with duplicates removed.
}
