% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadAstibble.R
\name{freadAstibble}
\alias{freadAstibble}
\title{Load the path using fread as a tibble}
\usage{
freadAstibble(path)
}
\arguments{
\item{path}{target path}
}
\description{
Load the path using fread as a tibble
}
